Simplify code and remove some code duplication
